+++ sdk/agentserver/Azure.AI.AgentServer.Core/tests/ServerVersionRegistryTests.cs
@@ -7,12 +7,12 @@ using NUnit.Framework;
 namespace Azure.AI.AgentServer.Core.Tests;
 
 [TestFixture]
-public class ServerUserAgentRegistryTests
+public class ServerVersionRegistryTests
 {
     [Test]
     public void Register_AddsSegment()
     {
-        var registry = new ServerUserAgentRegistry();
+        var registry = new ServerVersionRegistry();
         registry.Register("test-protocol/1.0.0");
         Assert.That(registry.GetSegments(), Has.Count.EqualTo(1));
         Assert.That(registry.GetSegments()[0], Is.EqualTo("test-protocol/1.0.0"));
@@ -21,7 +21,7 @@ public class ServerUserAgentRegistryTests
     [Test]
     public void Register_IgnoresDuplicates()
     {
-        var registry = new ServerUserAgentRegistry();
+        var registry = new ServerVersionRegistry();
         registry.Register("test-protocol/1.0.0");
         registry.Register("test-protocol/1.0.0");
         Assert.That(registry.GetSegments(), Has.Count.EqualTo(1));
@@ -30,7 +30,7 @@ public class ServerUserAgentRegistryTests
     [Test]
     public void Register_AllowsDistinctSegments()
     {
-        var registry = new ServerUserAgentRegistry();
+        var registry = new ServerVersionRegistry();
         registry.Register("protocol-a/1.0");
         registry.Register("protocol-b/2.0");
         Assert.That(registry.GetSegments(), Has.Count.EqualTo(2));
@@ -41,28 +41,28 @@ public class ServerUserAgentRegistryTests
     [Test]
     public void Register_ThrowsOnNull()
     {
-        var registry = new ServerUserAgentRegistry();
+        var registry = new ServerVersionRegistry();
         Assert.Throws<ArgumentNullException>(() => registry.Register(null!));
     }
 
     [Test]
     public void Register_ThrowsOnEmpty()
     {
-        var registry = new ServerUserAgentRegistry();
+        var registry = new ServerVersionRegistry();
         Assert.Throws<ArgumentException>(() => registry.Register(""));
     }
 
     [Test]
     public void GetSegments_ReturnsEmptyByDefault()
     {
-        var registry = new ServerUserAgentRegistry();
+        var registry = new ServerVersionRegistry();
         Assert.That(registry.GetSegments(), Is.Empty);
     }
 
     [Test]
     public void GetSegments_PreservesRegistrationOrder()
     {
-        var registry = new ServerUserAgentRegistry();
+        var registry = new ServerVersionRegistry();
         registry.Register("c/1.0");
         registry.Register("a/1.0");
         registry.Register("b/1.0");
@@ -75,7 +75,7 @@ public class ServerUserAgentRegistryTests
     [Test]
     public void GetSegments_ReturnsCopy()
     {
-        var registry = new ServerUserAgentRegistry();
+        var registry = new ServerVersionRegistry();
         registry.Register("x/1.0");
         var first = registry.GetSegments();
         registry.Register("y/2.0");
@@ -89,8 +89,8 @@ public class ServerUserAgentRegistryTests
     [Test]
     public void BuildIdentityString_FormatsCorrectly()
     {
-        var assembly = typeof(ServerUserAgentRegistry).Assembly;
-        var identity = ServerUserAgentRegistry.BuildIdentityString("test-sdk", assembly);
+        var assembly = typeof(ServerVersionRegistry).Assembly;
+        var identity = ServerVersionRegistry.BuildIdentityString("test-sdk", assembly);
         Assert.That(identity, Does.StartWith("test-sdk/"));
         Assert.That(identity, Does.Contain("(dotnet/"));
         Assert.That(identity, Does.EndWith(")"));
@@ -101,36 +101,36 @@ public class ServerUserAgentRegistryTests
     {
         // The assembly version may contain +commit-hash metadata
         // BuildIdentityString should strip everything after '+'
-        var assembly = typeof(ServerUserAgentRegistry).Assembly;
-        var identity = ServerUserAgentRegistry.BuildIdentityString("test-sdk", assembly);
+        var assembly = typeof(ServerVersionRegistry).Assembly;
+        var identity = ServerVersionRegistry.BuildIdentityString("test-sdk", assembly);
         Assert.That(identity, Does.Not.Contain("+"));
     }
 
     [Test]
     public void BuildIdentityString_ThrowsOnNullSdkName()
     {
-        var assembly = typeof(ServerUserAgentRegistry).Assembly;
-        Assert.Throws<ArgumentNullException>(() => ServerUserAgentRegistry.BuildIdentityString(null!, assembly));
+        var assembly = typeof(ServerVersionRegistry).Assembly;
+        Assert.Throws<ArgumentNullException>(() => ServerVersionRegistry.BuildIdentityString(null!, assembly));
     }
 
     [Test]
     public void BuildIdentityString_ThrowsOnEmptySdkName()
     {
-        var assembly = typeof(ServerUserAgentRegistry).Assembly;
-        Assert.Throws<ArgumentException>(() => ServerUserAgentRegistry.BuildIdentityString("", assembly));
+        var assembly = typeof(ServerVersionRegistry).Assembly;
+        Assert.Throws<ArgumentException>(() => ServerVersionRegistry.BuildIdentityString("", assembly));
     }
 
     [Test]
     public void BuildIdentityString_ThrowsOnNullAssembly()
     {
-        Assert.Throws<ArgumentNullException>(() => ServerUserAgentRegistry.BuildIdentityString("test-sdk", null!));
+        Assert.Throws<ArgumentNullException>(() => ServerVersionRegistry.BuildIdentityString("test-sdk", null!));
     }
 
     [Test]
     public void BuildIdentityString_IncludesRuntimeVersion()
     {
-        var assembly = typeof(ServerUserAgentRegistry).Assembly;
-        var identity = ServerUserAgentRegistry.BuildIdentityString("test-sdk", assembly);
+        var assembly = typeof(ServerVersionRegistry).Assembly;
+        var identity = ServerVersionRegistry.BuildIdentityString("test-sdk", assembly);
         var expected = $"dotnet/{Environment.Version.Major}.{Environment.Version.Minor}";
         Assert.That(identity, Does.Contain(expected));
     }